data-engineering/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
azure-core==1.30.1
azure-identity==1.16.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
azure-storage-queue==12.10.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
distro==1.9.0
Django==4.2.6
django-allauth==0.58.2
django-autoslug==1.9.9
django-cities-light==3.10.1
exceptiongroup==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
isodate==0.6.1
msal==1.28.0
msal-extensions==1.1.0
oauthlib==3.2.2
openai==1.35.10
packaging==24.0
pandas
Pillow==10.1.0
portalocker==2.8.2
progressbar2==4.4.2
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyodbc==5.1.0
pypyodbc==1.3.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-utils==3.8.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
SQLAlchemy==2.0.30
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stripe==6.7.0
tenacity==8.5.0
tqdm==4.66.4
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.8
urllib3
whitenoise==6.6.0
xero_python==3.0.0
